Code reorganization and code refactoring
Thie biggest change in this commit is that all of the code was reorganized into separate folders, and some code implementation (namely in the Compiler class) was split across multiple files. This was done for readability purposes.
Further, bugs were fixed in binary tree evaluation and in function definitions and calls; previously, it was not properly handling and tracking stack position, so any code that used the stack failed horribly upon being compiled and executed in the VM. The compiler now properly tracks and modifies stack offsets (including max offsets and where stack frames begin/end) so that this issue can be avoided.
See the commit history document for a mode detailed commit message.

+++ parser/Statement.cpp
@@ -16,12 +16,12 @@ stmt_type Statement::get_statement_type() {
 	return Statement::statement_type;
 }
 
-int Statement::get_line_number()
+unsigned int Statement::get_line_number()
 {
 	return this->line_number;
 }
 
-void Statement::set_line_number(int line_number)
+void Statement::set_line_number(unsigned int line_number)
 {
 	this->line_number = line_number;
 }
@@ -288,11 +288,11 @@ std::string Call::get_func_name() {
 	return this->func->getValue();
 }
 
-int Call::get_args_size() {
+size_t Call::get_args_size() {
 	return this->args.size();
 }
 
-std::shared_ptr<Expression> Call::get_arg(int num) {
+std::shared_ptr<Expression> Call::get_arg(size_t num) {
 	return this->args[num];
 }
